+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### 1.0.3 - 2014-12-10
+
+Fixes:
+- Random tabs don't trigger escapes.
+- Raw tags inside of code fences are properly stripped.
+- Double in-line code backticks are properly escaped.
+
 ### 1.0.2 - 2014-9-13
 - Fix: Removing extra raw tags in markdown code blocks.